minipager: new package for Paging functionality.
Split paging from miniclient into separate package so that we don't have
to import unsafe in miniclient. This allows us to remove the copied code
from minidoc.
