+++ src/minidoc/minimega.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"minicli"
+	"miniclient"
 	log "minilog"
 	"strings"
 )
@@ -13,7 +14,7 @@ func sendCommand(s string) string {
 
 	log.Debug("sendCommand: %v", s)
 
-	mm, err := Dial(*f_minimega)
+	mm, err := miniclient.Dial(*f_minimega)
 	if err != nil {
 		log.Errorln(err)
 		return err.Error()
